[IMP] sale_subscription: manage payments client side
Allow to de-structure the payment between frontend and backend
(notably in the case of Stripe) for SCA as part of the PSD2.

This means that instead of doing the transaction completely s2s in the
server, it is now possible to do it partially in the frontend according
to the following flow
1/ try to process the transaction in full s2s mode
2/ redirect to /payment/process
3/ if the tx is pending, ask for authentication client-side and reload
once done
4/ redirect to the return url of the tx (in this case the subscription)

What we gain: SCA flow for subscriptions
What we lose: confirmation message in the redirection following the tx

closes odoo/enterprise#5851

+++ sale_subscription/controllers/portal.py
@@ -7,6 +7,7 @@ from odoo import http
 from odoo.http import request
 from odoo.tools.translate import _
 
+from odoo.addons.payment.controllers.portal import PaymentProcessing
 from odoo.addons.portal.controllers.portal import get_records_pager, pager as portal_pager, CustomerPortal
 
 
@@ -176,7 +177,8 @@ class sale_subscription(http.Controller):
         if payment_token:
             invoice_values = account.sudo()._prepare_invoice()
             new_invoice = invoice_res.sudo().create(invoice_values)
-            tx = account.sudo()._do_payment(payment_token, new_invoice)[0]
+            tx = account.sudo().with_context(off_session=False)._do_payment(payment_token, new_invoice)[0]
+            PaymentProcessing.add_payment_transaction(tx)
             if tx.html_3ds:
                 return tx.html_3ds
             get_param = self.payment_succes_msg if tx.renewal_allowed else self.payment_fail_msg
@@ -184,10 +186,13 @@ class sale_subscription(http.Controller):
                 account.send_success_mail(tx, new_invoice)
                 msg_body = 'Manual payment succeeded. Payment reference: <a href=# data-oe-model=payment.transaction data-oe-id=%d>%s</a>; Amount: %s. Invoice <a href=# data-oe-model=account.move data-oe-id=%d>View Invoice</a>.' % (tx.id, tx.reference, tx.amount, new_invoice.id)
                 account.message_post(body=msg_body)
-            else:
+            elif tx.state != 'pending':
+                # a pending status might indicate that the customer has to authenticate, keep the invoice for post-processing
+                # NOTE: this might cause a lot of draft invoices to stay alive; i'm afraid this can't be helped
+                #       since the payment flow is divided in 2 in that case and the draft invoice must survive after the request
                 new_invoice.unlink()
 
-        return request.redirect('/my/subscription/%s/%s?%s' % (account.id, account.uuid, get_param))
+        return request.redirect('/payment/process')
 
     # 3DS controllers
     # transaction began as s2s but we receive a form reply
@@ -241,12 +246,8 @@ class sale_subscription(http.Controller):
 
         if kw.get('pm_id'):
             new_token = request.env['payment.token'].browse(int(kw.get('pm_id')))
-
-            if (new_token.verified and new_token.acquirer_id.check_validity) or not new_token.acquirer_id.check_validity:
-                account.payment_token_id = new_token
-                get_param = 'message=Your payment method has been changed for this subscription.&message_class=alert-success'
-            else:
-                get_param = 'message=Your payment method must be verified to use it on a subscription.&message_class=alert-danger'
+            account.payment_token_id = new_token
+            get_param = 'message=Your payment method has been changed for this subscription.&message_class=alert-success'
         else:
             get_param = 'message=Impossible to change your payment method for this subscription.&message_class=alert-danger'
